+++ public/index.html
@@ -45,7 +45,7 @@
       <!-- Контент -->
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
 
   </body>
 </html>
