+++ src/components/button-load-more.js
@@ -1,0 +1,5 @@
+/**
+ * Создание шаблона кнопки показа оставшихся задач
+ * @return {string} - кнопка
+ */
+export const createLoadMore = () => `<button class="load-more" type="button">load more</button>`;
