+++ src/components/consts.js
@@ -1,0 +1,11 @@
+const COUNT_TASKS = 3;
+
+const Position = {
+  BEFORE_END: `beforeend`,
+  AFTER_END: `afterend`
+};
+
+export {
+  COUNT_TASKS,
+  Position
+};
